backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyYAML
httpx
python-dotenv
openai
